refactor/#9: Analysis 디렉토리 Video로 병합
