+++ src/pages/Video/components/AnalyzeInfoCarousel.tsx
